Fixed #35, added report for each content
Teacher may now view the results of all students for a specifc
interactive content.
Note on the code: I was hoping we could have created some more
classes in core for the data view so that it would be easier
to use, however, we're a bit short on time now.

MDL-35