Add string for grade.php
Add string language to reference when there is no student submission when accessing the Grade report.

As of now, it is hardcoded inside grade.php ( $title = "Results for {$hvp->title}"; ). I propose to use this string with a conditional whenever there is no student submission, otherwise shows the default variable content.