Fix fetching shared content from the H5P hub
There is a bug in Moodle core (MDL-73588) that makes CURLOPT_FILE not
working as expected in cases when redirects are involved and the
redirecting pages actually generate some HTML output, too.

This bug affects H5P hub and it was not possible to fetch any shared
content from it in recent Moodle versions.

As a workaround, this patch stops using the CURLOPT_FILE option at all.